aider: Fix the "Inloggen" and "Account maken" links in the navbar to point to the correct routes.
# Aider chat conversation:

USER: de inloggen en account aanmaken links werken niet

ASSISTANT: To fix the "Inloggen" and "Account maken" links in the navbar, we need to update the URLs in the `resources\views\layouts\navbar.blade.php` file to point to the correct routes.

Here is the *SEARCH/REPLACE* block:

resources\views\layouts\navbar.blade.php
```php
<<<<<<< SEARCH
            <a href="#" class="hover:underline">Inloggen</a>
            <a href="#" class="hover:underline">Account maken</a>
=======
            <a href="{{ route('login') }}" class="hover:underline">Inloggen</a>
            <a href="{{ route('register') }}" class="hover:underline">Account maken</a>
>>>>>>> REPLACE
```

This will ensure that the "Inloggen" and "Account maken" links point to the correct login and register routes.